+++ src/main/java/com/coredisc/domain/OfficialQuestion.java
@@ -1,7 +1,7 @@
 package com.coredisc.domain;
 
 import com.coredisc.domain.common.BaseEntity;
-import com.coredisc.domain.mapping.QuestionCategory;
+import com.coredisc.domain.mapping.questionCategory.QuestionCategory;
 import com.coredisc.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.*;
