+++ src/main/java/com/coredisc/domain/category/CategoryRepository.java
@@ -1,0 +1,8 @@
+package com.coredisc.domain.category;
+
+import java.util.Optional;
+
+public interface CategoryRepository {
+
+    Optional<Category> findById(Long id);
+}
