+++ src/main/java/com/coredisc/domain/mapping/questionCategory/QuestionCategoryRepository.java
@@ -1,0 +1,6 @@
+package com.coredisc.domain.mapping.questionCategory;
+
+public interface QuestionCategoryRepository {
+
+    QuestionCategory save(QuestionCategory questionCategory);
+}
