+++ src/main/java/com/coredisc/domain/personalQuestion/PersonalQuestionRepository.java
@@ -1,0 +1,6 @@
+package com.coredisc.domain.personalQuestion;
+
+public interface PersonalQuestionRepository {
+
+    PersonalQuestion save(PersonalQuestion personalQuestion);
+}
